Add files via upload
Here is the wok in progress as of Nov 1st 2018. I have been working on it for 4 hours. I have a lot left to finish. Including building/converting it to a wordpress theme but as it is Friday tomorrow I thought I would be best to show you where I am at with the project. 

It has been difficult to focus at 100% capacity with working 9-5 followed by heavy traffic and an office move this week.  I broke the DRY rule a few times in my sass, the overall look of the page that I have completed is responsive to tablet (Once the page is finished I will begin cleaning up the sass, adding the hamburger menu for mobile devices, altering paddings/margins and font sizes etc to be closer to pixel perfect with the spec).

I have enjoyed this project and if I have the weekend to focus my full attention on it I will have a finished pixel perfect page with some jquery for the search bar show/hide and a wordpress theme of it.

Kind regards
J